对获取GitHub profile 部分进行代码重构
