+++ dags/libs/base_dict/opensearch_index.py
@@ -1,9 +1,8 @@
 OPENSEARCH_GIT_RAW = "git_raw"
 OPENSEARCH_GIT_GITHUB_CLEAN = "git_github_clean"
-
 OPENSEARCH_INDEX_GITHUB_COMMITS = "github_commits"
-
 OPENSEARCH_INDEX_GITHUB_ISSUES = "github_issues"
 OPENSEARCH_INDEX_GITHUB_ISSUES_COMMENTS = "github_issues_comments"
-
 OPENSEARCH_INDEX_CHECK_SYNC_DATA = "check_sync_data"
+OPEN_SEARCH_GITHUB_PROFILE_INDEX = "github_profile"
+OPENSEARCH_INDEX_GITHUB_ISSUES_TIMELINE = "github_issues_timeline"
