+++ dags/libs/github/sync_profiles.py
@@ -1,7 +1,9 @@
 import itertools
-
 from loguru import logger
-from . import init_profile_commen
+from . import init_profiles
+import requests
+from ..base_dict.opensearch_index import OPEN_SEARCH_GITHUB_PROFILE_INDEX
+from ..util.github_api import GithubAPI
 
 
 # TODO: 传入用户的profile信息，获取用户的location、company、email，与晨琪对接
@@ -28,13 +30,13 @@ def github_profile_data_source(now_github_profile):
 
 # TODO: 定时任务更新github_profile
 def add_updated_github_profiles(github_tokens, opensearch_conn_infos):
-    opensearch_client = init_profile_commen.get_opensearch_client(opensearch_conn_infos)
+    opensearch_client = init_profiles.get_opensearch_client(opensearch_conn_infos)
 
     github_tokens_iter = itertools.cycle(github_tokens)
 
     # 将折叠（collapse）查询opensearch（去重排序）的结果作为查询更新的数据源
     existing_github_profiles = opensearch_client.search(
-        index='github_profile',
+        index=OPEN_SEARCH_GITHUB_PROFILE_INDEX,
         body={
             "query": {
                 "match_all": {}
@@ -61,12 +63,12 @@ def add_updated_github_profiles(github_tokens, opensearch_conn_infos):
         existing_github_profile_user_updated_at = existing_github_profile["_source"]["updated_at"]
 
         # 根据上述"login"信息获取git上的profile信息中的"updated_at1"信息
-        existing_github_profile_login = existing_github_profile["_source"]["login"]
-        now_github_profile = init_profile_commen.get_github_profile(github_tokens_iter, existing_github_profile_login,
-                                                                    opensearch_conn_infos)
+        existing_github_login = existing_github_profile["_source"]["login"]
+        github_api = GithubAPI()
+        session = requests.Session()
+        now_github_profile = github_api.get_github_profiles(http_session=session, github_tokens_iter=github_tokens_iter, login_info=existing_github_login)
         # todo: test -->delete
         github_profile_data_source(now_github_profile)
-
         now_github_profile = json.dumps(now_github_profile)
         now_github_profile_user_updated_at = json.loads(now_github_profile)["updated_at"]
 
@@ -74,7 +76,7 @@ def add_updated_github_profiles(github_tokens, opensearch_conn_infos):
         # 一致：不作处理
         # 不一致：将新的profile信息添加到OpenSearch中
         if existing_github_profile_user_updated_at != now_github_profile_user_updated_at:
-            opensearch_client.index(index="github_profile",
+            opensearch_client.index(index=OPEN_SEARCH_GITHUB_PROFILE_INDEX,
                                     body=now_github_profile,
                                     refresh=True)
             logger.info("Put the github user's new profile into opensearch.")
