+++ dags/libs/util/github_api.py
@@ -37,3 +37,24 @@ class GithubAPI:
         logger.info(f"url:{url}, \n headers:{headers}, \n params：{params}")
 
         return res
+
+    def get_github_profiles(self, http_session, github_tokens_iter, login_info):
+        """Get GitHub user's latest profile from GitHUb."""
+        url = "https://api.github.com/users/{login_info}".format(
+            login_info=login_info)
+        headers = copy.deepcopy(self.github_headers)
+        headers.update({'Authorization': 'token %s' % next(github_tokens_iter),
+                               'user-agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) '
+                                             'Chrome/96.0.4664.110 Safari/537.36'})
+        headers.update({'Authorization': 'token %s' % next(github_tokens_iter)})
+        params = {}
+        now_github_profile = {}
+        try:
+            req = do_get_result(http_session, url, headers, params)
+            now_github_profile = req.json()
+        except TypeError as e:
+            print("捕获airflow抛出的TypeError:", e)
+        logger.info(self.get_github_profiles.__doc__)
+        return now_github_profile
+
+
